// packages/frontend/src/generated/types/TruckWaybillCompositeRequest.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { TruckWaybillItemCompositeRequest } from './TruckWaybillItemCompositeRequest.ts'
import type { TruckWeightDocCompositeRequest } from './TruckWeightDocCompositeRequest.ts'

export interface TruckWaybillCompositeRequest {
  /**
   * @type string, date
   */
  date: string
  /**
   * @type string
   */
  documentNumber: string
  /**
   * @type array,null
   */
  items?: TruckWaybillItemCompositeRequest[] | null
  /**
   * @type string, uuid
   */
  senderId: string
  /**
   * @type array,null
   */
  weightDocs?: TruckWeightDocCompositeRequest[] | null
}