// packages/frontend/src/generated/zod/truckWaybillCompositeRequestSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ToZod } from '../.kubb/ToZod.ts'
import type { TruckWaybillCompositeRequest } from '../types/TruckWaybillCompositeRequest.ts'
import * as z from 'zod'
import { truckWaybillItemCompositeRequestSchema } from './truckWaybillItemCompositeRequestSchema.ts'
import { truckWeightDocCompositeRequestSchema } from './truckWeightDocCompositeRequestSchema.ts'

export const truckWaybillCompositeRequestSchema = z.object({
  date: z.string().date(),
  documentNumber: z.string(),
  items: z
    .array(z.lazy(() => truckWaybillItemCompositeRequestSchema))
    .nullish(),
  senderId: z.string().uuid(),
  weightDocs: z
    .array(z.lazy(() => truckWeightDocCompositeRequestSchema))
    .nullish(),
}) as unknown as ToZod<TruckWaybillCompositeRequest>