// packages/frontend/src/generated/zod/truckWaybillItemCompositeRequestSchema.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ToZod } from '../.kubb/ToZod.ts'
import type { TruckWaybillItemCompositeRequest } from '../types/TruckWaybillItemCompositeRequest.ts'
import * as z from 'zod'

export const truckWaybillItemCompositeRequestSchema = z.object({
  declaredAmount: z.string(),
  productId: z.string().uuid(),
}) as unknown as ToZod<TruckWaybillItemCompositeRequest>